fixed list BUG

+++ Data-Structures-Ex2/Bridge.h
@@ -8,8 +8,8 @@ private:
 
 public:
 	Bridge(float height = 0, Bridge* next = nullptr);
-	void setHeight(int height) { _height = height; }
-	int getHeight() { return _height; }
+	void setHeight(float height) { _height = height; }
+	float getHeight() { return _height; }
 	void setNext(Bridge* next) { _next = next; }
 	Bridge* getNext() { return _next; }
 };
